+++ src/WreckingBalloon/balloon.h
@@ -5,11 +5,11 @@
 
 typedef struct {
     Sprite *sprite;
-    uint    points;
+    uint16    points;
     bool    popped;
 } PointBalloon;
 
-void balloon_init(PointBalloon *pb, Sprite *sprite, uint points);
+void balloon_init(PointBalloon *pb, Sprite *sprite, uint16 points);
 void balloon_pop(PointBalloon *pb);
 void balloon_update(PointBalloon *pb);
